Switch:js>php. Add:vid info to database,embed vids
Switched from javascript to php for added security for mysql queries. Added ability for new video IDs and titles associated with breed results to be added to a database table. Added ability to check if video ID/title is already in table to ensure there are no duplicates. Based off results and videoIDs, a corresponding YouTube video will now appear on results page based off what is saved in database.