Error Debugging
So I did some error debugging. Things that changed:

If the YouTube API reaches its limit, then it'll skip the API and just pull from the saved videos in the database. I had to add an extra breed column to vidPups to make this work. But now it means I don't have to keep changing the key and if someone visits the site after we reach the quota they won't see ugly errors. A future problem could be if we add a new dog and the quota is reached so it can't grab a new video but I don't think we're going to add a breed so it should be okay. 

If there's a special character anywhere (dog name, shop name, address, city, etc) there shouldn't be a problem now for it to be inserted into the tables. Due to that, I had to change the code for how to display the petfinder results but it displays in the same way as before. 

I reduced the 100 radius for petfinder results to 50 since we debugged before how to show 0-3 dogs and I changed the pet shop radius to 30 miles. I'm thinking of adding a little drop down menu in puppypicker so they can choose the radius for how far the pups should be. 

I think that's it.